First commit!
This is actually a very old project I had, sitting on my hard drive since december 2014.
The goal was to implement all (hum, maybe not, so let's say as much as possible will do!) the commands defined by the lastest specification released by the Open Group (which is the issue 7 since the beginning of the project up to today at least).

At that time Rust basically broke all the time, and since I was trying to learn it, it was kind of... annoying.
The latest version of Rust Rustine ran on was 0.13 and I stopped working on it at that time.
But today I've found it and have decided to fix it. And well... I have basically reimplemented everything except pwd (which I of course had to fix though), because the code I used to be proud of was kinda bad.
So yeah, now the quality should be okay (exept for pwd maybe), even though there are not many utilities implemented.

My motivation for fixing this is that I am creating a shell, Rush, which will need as many builtins as possible at some point, so these could come in handy!